OBPIH-6532 Replace ActionDots with new component ContextMenu
