+++ src/js/consts/actionItemType.js
@@ -1,0 +1,7 @@
+const actionItemType = {
+  LINK: 'LINK',
+  REACT_LINK: 'REACT_LINK',
+  BUTTON: 'BUTTON',
+};
+
+export default actionItemType;
