+++ src/js/tests/utils/ActionDots.test.jsx
@@ -1,9 +1,9 @@
 import React from 'react';
 
-import { fireEvent, render, screen, waitFor } from '@testing-library/react';
+import { fireEvent, render, screen } from '@testing-library/react';
 import renderer from 'react-test-renderer';
 
-import ActionDots from 'utils/ActionDots';
+import ContextMenu from 'utils/ContextMenu';
 
 import '@testing-library/jest-dom';
 
@@ -32,7 +32,7 @@ describe('action dots', () => {
       defaultLabel: 'defaultLabelTest',
       leftIcon: <React.Fragment>leftIconTest</React.Fragment>,
     }];
-    actionDots = (<ActionDots
+    actionDots = (<ContextMenu
       id="1"
       actions={actions}
     />);
@@ -58,32 +58,5 @@ describe('action dots', () => {
     fireEvent.click(screen.getByTestId(testId));
     expect(screen.findByTestId(testId))
       .toBeTruthy();
-  });
-
-  it('should have hidden dropdown-menu', () => {
-    render(actionDots);
-    expect(screen.queryByTestId('dropdown-toggle'))
-      .toBeTruthy();
-    expect(screen.getByTestId('dropdown-menu'))
-      .toHaveStyle({ display: 'none' });
-  });
-
-  it('should have visible dropdown-menu', async () => {
-    const testId = 'dropdown-menu';
-    const { rerender } = render(actionDots);
-    expect(screen.queryByTestId(testId))
-      .toHaveStyle({ display: 'none' });
-    fireEvent.click(screen.getByTestId(testId));
-    rerender(<ActionDots
-      id="1"
-      actions={actions}
-      dropdownClasses="show"
-    />);
-    // after rerender, previously injected styles are deleted, so I had to include it one more time.
-    applyStyles();
-    await waitFor(() => {
-      expect(screen.queryByTestId(testId))
-        .toHaveStyle({ display: 'block' });
-    });
   });
 });
